Added email sending to the duplicates report script, refactored emailing for background tasks so it can be shared with provenance reporting.
